+++ Kernel/System/SystemMaintenance.pm
@@ -512,13 +512,16 @@ sub SystemMaintenanceIsActive {
 
 =head2 SystemMaintenanceIsComing()
 
-get a SystemMaintenance flag
+Get a upcoming SystemMaintenance start and stop date.
 
-    my $SystemMaintenanceIsComing = $SystemMaintenanceObject->SystemMaintenanceIsComing();
+    my %SystemMaintenanceIsComing = $SystemMaintenanceObject->SystemMaintenanceIsComing();
 
 Returns:
 
-    $SystemMaintenanceIsComing = 1 # 1 or 0
+    %SystemMaintenanceIsComing = {
+        StartDate => 1515614400,
+        StopDate  => 1515607200
+    };
 
 =cut
 
@@ -538,7 +541,7 @@ sub SystemMaintenanceIsComing {
     my $DBObject = $Kernel::OM->Get('Kernel::System::DB');
 
     my $SQL = "
-            SELECT start_date
+            SELECT start_date, stop_date
             FROM system_maintenance
             WHERE start_date > $SystemTime and start_date <= $TargetTime
     ";
@@ -554,16 +557,17 @@ sub SystemMaintenanceIsComing {
 
     return if !$DBObject->Prepare( SQL => $SQL );
 
-    my $Result;
+    my %Result;
     RESULT:
     while ( my @Row = $DBObject->FetchrowArray() ) {
-        $Result = $Row[0];
+        $Result{StartDate} = $Row[0];
+        $Result{StopDate}  = $Row[1];
         last RESULT;
     }
 
-    return if !$Result;
+    return if !%Result;
 
-    return $Result;
+    return %Result;
 }
 
 1;